requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank_bm25==0.2.2
torch==2.0.1
transformers==4.30.2
streamlit==1.28.1